backend/requirements.txt
abydos==0.5.0
alembic==1.9.2
annotated-types==0.6.0
Authlib==1.3.0
autopep8==1.7.0
blinker==1.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cryptography==38.0.3
cycler==0.11.0
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
docx==0.2.4
ecdsa==0.18.0
Flask==3.0.0
Flask-Alembic==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.1
flask-oidc==2.1.1
Flask-pyoidc==3.14.3
Flask-SQLAlchemy==3.1.1
fonttools==4.39.3
future==0.18.3
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Levenshtein==0.21.1
lxml==4.9.1
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mypy==0.991
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.3
oic==1.6.1
packaging==23.1
Pillow==9.2.0
pipreqs==0.4.11
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyjwkest==1.4.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==3.7.0
python-Levenshtein==0.21.1
rapidfuzz==3.3.0
regex==2023.5.5
requests==2.28.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.21
sqlalchemy-stubs==0.4
textdistance==4.5.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.10.0
urllib3==1.26.12
Werkzeug==3.0.0
wordcloud==1.9.1.1
yarg==0.1.9
psutil==7.0.0
